+++ src/lib/components/private-menu/index.vue
@@ -247,7 +247,7 @@ export default {
             immediate : true,
             deep : true
         });
-
+        
     }
 };
 </script>